Bind onTest to an instance of Test
Since onTest is being called before each test people use it to not
only generate test arguments but to add additional properties to
the test case and so on. This patch makes the code below possible:

  hiro.module("MySuite", {
		onTest: function () {
			this.doctor =  "who";
		},

		testSomething: function () {
			myFunc(this.doctor); // Will pass "who"
		}
	});

However, to think about it, we just need to add an ability to
have custom properties and methods in the suite declaration but
that's an after-release task.