Change in-test context to a mixin of Test and Suite
Apparently, you need to have methods from both Test and Suite
instances inside of your test cases. Due to backwards compatibility
it has to be bound to 'this', we can't just introduce a new object
or argument.

That's why we have to mixing Test and Suite objects together
when test case is executed as a part of a suite (as opposed to
manually creating and running a single Test instance).