Add ability to run only one suite
When working on a feature it is often nice to be able to run one
particular suite. This patch modifies webui so that it reacts to
a suite name being a part of a query string:

  http://localhost:7777/ <-- runs all tests
  http://localhost:7777/BasicTests <-- runs only BasicTests

There's no UI for this yet because I'm not sure what's the best
solution.

Even though the solution is partial it closes GH-23.